Add warnings to VFs in todo list, share sample text component
